Update test help printing and BUILD
Added a work around for the TEST help printing code.  Added a script to auto-generate a build number.

+++ src/cli/test_cmds.h
@@ -8,6 +8,6 @@
 /*
  * Test command
  */
-extern short cmd_test(short channel, int argc, char * argv[]);
+extern int cmd_test(short channel, int argc, char * argv[]);
 
 #endif
